app/pyproject.toml
[tool.poetry]
name = "mixel-music"
version = "0.3.1"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.30.6"
fastapi = "^0.115.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
pydantic-settings = "^2.5.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.35"
watchfiles = "^0.24.0"
aiosqlite = "^0.20.0"
aiofiles = "^24.1.0"
pillow = "^10.4.0"
rich = "^13.8.1"
toml = "^0.10.2"
tinytag = "^2.0.0"
diskcache = "^5.6.3"
itsdangerous = "^2.2.0"
argon2-cffi = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
